feat: define and auto-generate the data model
Here the idea is that we have a master data model inside of the
internal/datamodel folder. This data model will be used to generate
the client along with the apimodel (not yet implemented).

The data model will also be copied varbatim as a generated file
into the top-level folder. This way, consumers of this package will
not need to be concerned about implementation details.