feat: implement TorTargets API
While there, acknowledge that there are some APIs that return
data types different from data structures. So, make sure we can
actually define such APIs. For now, the only APIs with these
characteristics are PsiphonConfig and TorTargets.

We probably need to do some more hammering to make working with
such APIs a bit more natural than it is now. For example, it would
be very cool to avoid returning a pointer to a map.